fix input and inline diff
